// production/src/home/popup-type-enum.ts
export enum PopupTypeEnum {
  dataPreparation = 'data-preparation',
  editFilters = 'edit-filters',
  loadingPage = 'loading-page',
  refreshAllPage = 'refresh-all-page',
  promptsWindow = 'prompts-window',
  repromptingWindow = 'reprompting-window',
  libraryWindow = 'library-window',
  dossierWindow = 'dossier-window',
  obtainInstanceHelper = 'obtain-instance-helper',
  multipleRepromptTransitionPage = 'multiple-reprompt-transition-page',
  importedDataOverview = 'imported-data-overview',
  repromptDossierDataOverview = 'reprompt-dossier-data-overview',
  repromptReportDataOverview = 'reprompt-report-data-overview',
}